PO_DBXref.txt: Added stanza for OB_SF_PO (used to link to SF trackers) I changed the time and date stamp manually.
poc_project.html:
Put “the” in the title.
Changed Plant Structure to Plant Anatomy and reworded first lines. Provided more current examples.
Changed wording throughout to say “is being developed” instead of “will be developed” and similar.
Replaced explanations of relations with link to relations page on the wiki.
Removed the section on syntax. It was out of date and irrelevant.
Removed the section on use of symbols in flat files.
Removed the section on Domain Conceptualization. It is not clear what it is trying to say or why, and, to the extent I can understand it, it does not reflect how we are currently developing the ontology.